moved the barcode scanner code into another component
